backend/requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
numpy==2.0.1
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.5.1
scipy==1.13.1
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.1
Werkzeug==3.0.3
zipp==3.19.2
